services_ML/requirements.txt
# FastAPI y servidor
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Base de datos PostgreSQL
psycopg[binary]==3.2.3
psycopg[pool]==3.2.3
pgvector>=0.2.5,<0.3.0

# Machine Learning - Clustering
hdbscan>=0.8.40  # Usar version precompilada para Windows
scikit-learn==1.5.2

# Machine Learning - Reducción Dimensional
umap-learn==0.5.6

# Machine Learning - Topic Modeling
bertopic==0.16.4

# Embeddings (compartido con services_LLM)
sentence-transformers==3.3.1
huggingface-hub==0.26.2

# Análisis de datos
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1

# Visualización
plotly==5.24.1
matplotlib==3.9.2

# Utilidades
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1

# Logging
loguru==0.7.3

# Testing (opcional)
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
